Squashed commit of the following:
commit 779d9f318acd9983ccc66cb4f135a1e2fc3ad249
Merge: 4b3262b 0d6d201
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Mon Oct 2 21:56:44 2023 +0800

    Merge branch 'feat-nolog3' into deployment

commit 0d6d2013260f5714bae5e8364bd8a09f428e6485
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Mon Oct 2 21:53:15 2023 +0800

    remove log

commit 07feaa9af8fb6dcfef96d49682e3847f2a50fba0
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Mon Oct 2 21:42:29 2023 +0800

    adjust tests

commit 3cff3ff2239182f1a14b76858dbb4b471819ddf0
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Mon Sep 25 17:05:00 2023 +0800

    fix oracle

commit 4b3262b65236cf30957a247bc62977d7b9368fa5
Merge: 94d770c 28dd988
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Sun Sep 24 22:25:17 2023 +0800

    Merge branch 'feat-nolog2' into deployment

commit 28dd9887bcdb8dd372fa7ca20a79d89bc2195ddf
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Sun Sep 24 22:12:34 2023 +0800

    rm log

commit e29267b26fa7a7bc35a55615e90c8a753616d00f
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Sun Sep 24 22:05:12 2023 +0800

    adjust test

commit 94d770c4bcf8b96ce627f641a2450554d20a3cb8
Merge: 5573abb 632fb13
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Fri Sep 22 21:42:17 2023 +0800

    Merge branch 'feat-nolog1' into deployment

commit 632fb1357a6533e9749de826efd5503df6d7da52
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Fri Sep 22 21:40:08 2023 +0800

    remove log

commit 2df7e857e9e1669de7807287d4675120d89d6f23
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Fri Sep 22 21:27:22 2023 +0800

    fix addLiquidity

commit 0e13bb5626e4fb7f642922c247f3959f0e3355d9
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Fri Sep 22 18:18:02 2023 +0800

    handle warnnings

commit dfff96512904641d16e5f1b63d3f6e7f1d2c0e00
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Wed Sep 20 17:08:56 2023 +0800

    fix oracle

commit 3f7e689758d786403f0e5f189a7e8a6726e9874e
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Tue Sep 19 17:58:38 2023 +0800

    remove some dead code

commit 5573abb8e4a80ecfd69a999450f1899c3f68df03
Merge: 8c2f516 723cb47
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Tue Sep 19 16:37:02 2023 +0800

    Merge branch 'feat-nolog' into deployment

commit 723cb477a1f6f69a59e7ec4c60dd756b428df7a9
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Tue Sep 19 16:31:32 2023 +0800

    remove log

commit 7922b14b2d6d26af334035f3943380dfb895730e
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Tue Sep 19 16:25:19 2023 +0800

    add log

commit bec0d6574aa28ffd6a2b8a8ad0c3282cb5743121
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Tue Sep 19 16:12:09 2023 +0800

    fix TradeMath

commit dbddeca1d777ea388a2a1da64835c97ec95b30b3
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Mon Sep 18 15:55:53 2023 +0800

    update

commit 8c2f516e03f2064a2f27cceb2850e94c173627f3
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Thu Sep 14 18:19:36 2023 +0800

    ready

commit 2564c02e9d7304aa8624c871bdcc7525dd33de16
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Thu Sep 14 18:12:54 2023 +0800

    more tests

commit 16e4e137bf50a4c48e3feac23d334ade2d4f0528
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Mon Aug 28 10:59:09 2023 +0800

    alter oracle

commit 38056fe7dcd2e8d8e4014d37645301141d63cf56
Author: Alexis-ETH <teeterbritnitm43@gmail.com>
Date:   Wed Aug 23 10:54:30 2023 +0800

    for debug

+++ src/core/interfaces/battle/IBattleState.sol
@@ -2,7 +2,7 @@
 
 pragma solidity ^0.8.0;
 
-import "../../types/common.sol";
+import "core/types/common.sol";
 
 interface IBattleState {
     /// @notice Get slotInfo by slotKey.
@@ -37,5 +37,5 @@ interface IBattleState {
 
     function getInsideLast(int24 tickLower, int24 tickUpper) external view returns (GrowthX128 memory);
 
-    function fee() external view returns(uint, uint , uint);
+    function fee() external view returns (uint256, uint256, uint256);
 }
